requirements.txt
click==6.7
dnspython==1.15.0
eventlet==0.24.1
Flask==1.0.2
Flask-SocketIO==3.0.2
greenlet==0.4.15
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
monotonic==1.5
pkg-resources==0.0.0
PyCmdMessenger==0.2.4
pyserial==3.4
python-engineio==2.2.0
python-socketio==2.0.0
six==1.11.0
Werkzeug==0.14.1
